[PM-6702] Move get and store urls for email from auth disk source to environment disk source
