+++ app/src/test/java/com/x8bit/bitwarden/data/auth/datasource/disk/util/FakeAuthDiskSource.kt
@@ -2,7 +2,6 @@ package com.x8bit.bitwarden.data.auth.datasource.disk.util
 
 import com.x8bit.bitwarden.data.auth.datasource.disk.AuthDiskSource
 import com.x8bit.bitwarden.data.auth.datasource.disk.model.AccountTokensJson
-import com.x8bit.bitwarden.data.auth.datasource.disk.model.EnvironmentUrlDataJson
 import com.x8bit.bitwarden.data.auth.datasource.disk.model.PendingAuthRequestJson
 import com.x8bit.bitwarden.data.auth.datasource.disk.model.UserStateJson
 import com.x8bit.bitwarden.data.platform.repository.util.bufferedMutableSharedFlow
@@ -44,7 +43,6 @@ class FakeAuthDiskSource : AuthDiskSource {
     private val storedBiometricKeys = mutableMapOf<String, String?>()
     private val storedMasterPasswordHashes = mutableMapOf<String, String?>()
     private val storedPolicies = mutableMapOf<String, List<SyncResponseJson.Policy>?>()
-    private val storedEmailVerificationUrls = mutableMapOf<String, EnvironmentUrlDataJson?>()
 
     override var userState: UserStateJson? = null
         set(value) {
@@ -214,13 +212,6 @@ class FakeAuthDiskSource : AuthDiskSource {
     override fun storeAccountTokens(userId: String, accountTokens: AccountTokensJson?) {
         storedAccountTokens[userId] = accountTokens
         getMutableAccountTokensFlow(userId = userId).tryEmit(accountTokens)
-    }
-
-    override fun getEmailVerificationUrls(userEmail: String): EnvironmentUrlDataJson? =
-        storedEmailVerificationUrls[userEmail]
-
-    override fun storeEmailVerificationUrls(userEmail: String, urls: EnvironmentUrlDataJson) {
-        storedEmailVerificationUrls[userEmail] = urls
     }
 
     /**
